+++ plugin-boilerplate/{{ cookiecutter.project_slug }}/plugins/{{ cookiecutter.project_module }}/web-frontend/plugins.js
@@ -1,0 +1,7 @@
+import { BaserowPlugin } from '@baserow/modules/core/plugins'
+
+export class PluginNamePlugin extends BaserowPlugin {
+  static getType() {
+    return '{{ cookiecutter.project_slug }}'
+  }
+}