+++ changelog.md
@@ -12,8 +12,9 @@
 * Fixed keeping the datepicker visible in the grid view when selecting a date for the 
   first time.
 * Improved API 404 errors by providing a machine readable error.
+* Added cookiecutter plugin boilerplate.
 
-## Released (2020-08-06)
+## Released (2020-06-08)
 
 * Fixed not handling 500 errors.
 * Prevent row context menu when right clicking on a field that's being edited.